Recover cluster state from database
In the current cluster implementation all the root blocks persisted by
the root chain are guaranteed to have themselves and their included
minor blocks persisted by all the shards.

Based on this invariant it's easier to recover the cluster state based
on the root chain than allowing each shard to recover their own states
and then consolidating the states with each other. This also aligns with
the root chain first consensus protocol.

The recovery procedure includes the following steps.
1. Master recovers the root chain state by traversing back from the tip hash
2. Master sends the root tip over the Ping message when connecting to slaves
3. Slaves recovers the shard states to the most recent minor headers
included in the root tip