New Transaction data structure
There are 3 types of transactions:
1. Value transfer. In-shard transaction if fromFullShardId and toFullShardId
refer to the same shard, otherwise it is a cross-shard transaction.

2. Contract creation. 'to' must be empty. fromFullShardId and toFullShardId
must refer to the same shard id. The contract address will have the same
full shard id as toFullShardId. If the contract does not invoke other contract
normally the toFullShardId should be the same as fromFullShardId.

3. Contract call. fromFullShardId and toFullShardId must refer to the same
shard id based on the current number of shards in the network. It is possible
a reshard event would invalidate a tx that was valid before the reshard.